// src/Model/FlightBooking.ts
export class FlightBooking
{
    bookingId: string;
    passengerName: string;
    age: number;
    seatsBooked: number;
    journeyDate:Date;
    journeyTime:Date;
    flightId: string;
    totalAmount:number;

    // constructor(bookingId:string, passengerName:string, age:number, seatsBooked:number, flightId:string)
    // {
    //     this.bookingId=bookingId;
    //     this.passengerName=passengerName;
    //     this.age=age;
    //     this.seatsBooked=seatsBooked;
    //     this.flightId=flightId;
    // }
}